pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=7"]
build-backend = "setuptools.build_meta"

[project]
name = "fast_track"
version = "0.10.0"
description = "Object detection and tracking pipeline"
readme = "README.md"
keywords = [
    "computer-vision",
    "object-tracking",
    "object-detection",
    "ai",
]
license = { file = "LICENSE.txt" }
authors = [{ name = "Nate Haddad", email = "nhaddad2112@gmail.com" }]
maintainers = [{ name = "Nate Haddad", email = "nhaddad2112@gmail.com" }]
requires-python = ">=3.8, <3.11"
dependencies = [
    "numpy == 1.23.0", 
    "pyyaml == 6.0.1",
    "opencv-python == 4.9.0.80",
    "onnxruntime == 1.13.1",
    "torch == 2.2.0",
    "scipy == 1.12.0",
    "ultralytics == 8.1.9",
    "super-gradients == 3.6.0",
    "pylint == 3.0.3",
]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu==1.17.0",
]
bytetrack = [
    "lap == 0.4.0",
    "Cython == 3.0.8",
    "cython-bbox @ git+https://github.com/samson-wang/cython_bbox.git#egg=cython-bbox",
]
openaianalyst = [
    "langchain == 0.1.7",
    "langchain-community == 0.0.20",
    "langchain-experimental == 0.0.51",
    "langchain-openai == 0.0.6",
    "python-dotenv == 1.0.1",
]
gradio = [
    "gradio == 3.40.1",
]

[tool.setuptools]
py-modules = []
